requirements.txt
pycryptodome==3.15.0
loguru==0.5.3